+++ src/components/render/Properties.tsx
@@ -1,21 +1,26 @@
 import toTitleCase from 'to-title-case';
+import SettingsContext from '../../SettingsContext';
 
 export default function Properties(props: PropertiesProps) {
   return (
-    <div className="columns">
-      <div className="column is-narrow">
-        {Object.keys(props.value).map((key) => (
-          <div className="has-text-right has-text-weight-bold" key={key}>
-            {toTitleCase(key)}
+    <SettingsContext.Consumer>
+      {({ titleCasePropertyNames }) => (
+        <div className="columns">
+          <div className="column is-narrow">
+            {Object.keys(props.value).map((key) => (
+              <div className="has-text-right has-text-weight-bold" key={key}>
+                {titleCasePropertyNames ? toTitleCase(key) : key}
+              </div>
+            ))}
           </div>
-        ))}
-      </div>
-      <div className="column">
-        {Object.values(props.value).map((value, index) => (
-          <div key={index}>{String(value)}</div>
-        ))}
-      </div>
-    </div>
+          <div className="column">
+            {Object.values(props.value).map((value, index) => (
+              <div key={index}>{String(value)}</div>
+            ))}
+          </div>
+        </div>
+      )}
+    </SettingsContext.Consumer>
   );
 }