+++ src/components/settings/SettingsModal.tsx
@@ -1,0 +1,21 @@
+import {
+  ModalCard,
+  ModalCardBody,
+  ModalCardFoot,
+  ModalProps
+} from '../ModalCard';
+import HeadersSetting from './HeadersSetting';
+import TitleCasePropertyNamesSetting from './TitleCasePropertyNamesSetting';
+
+export default function SettingsModal(props: ModalProps) {
+  return (
+    <ModalCard {...props} title="Settings">
+      <ModalCardBody>
+        <TitleCasePropertyNamesSetting />
+        <hr />
+        <HeadersSetting />
+      </ModalCardBody>
+      <ModalCardFoot />
+    </ModalCard>
+  );
+}